+++ src/async_monitored_item.rs
@@ -6,8 +6,9 @@ use tokio::sync::mpsc::{self, error::TrySendError};
 use crate::{
     attributes, create_monitored_items, delete_monitored_items,
     monitored_item::{DataChange, Unknown},
-    ua, AsyncSubscription, DataType as _, Error, MonitoredItemAttribute, MonitoredItemHandle,
-    MonitoredItemKind, MonitoringFilter, Result,
+    ua, AsyncSubscription, DataType as _, Error, MonitoredItemAttribute,
+    MonitoredItemCreateRequestBuilder, MonitoredItemHandle, MonitoredItemKind, MonitoringFilter,
+    Result,
 };
 
 /// Maximum number of buffered values.
@@ -16,15 +17,15 @@ const DEFAULT_STREAM_BUFFER_SIZE: usize = 3;
 
 #[derive(Debug)]
 pub struct MonitoredItemBuilder<K: MonitoredItemKind> {
-    node_ids: Vec<ua::NodeId>,
-    attribute_id: ua::AttributeId,
-    monitoring_mode: Option<ua::MonitoringMode>,
+    pub(crate) node_ids: Vec<ua::NodeId>,
+    pub(crate) attribute_id: ua::AttributeId,
+    pub(crate) monitoring_mode: Option<ua::MonitoringMode>,
     #[expect(clippy::option_option, reason = "implied default vs. unset")]
-    sampling_interval: Option<Option<Duration>>,
-    filter: Option<Box<dyn MonitoringFilter>>,
-    queue_size: Option<u32>,
-    discard_oldest: Option<bool>,
-    _kind: PhantomData<K>,
+    pub(crate) sampling_interval: Option<Option<Duration>>,
+    pub(crate) filter: Option<Box<dyn MonitoringFilter>>,
+    pub(crate) queue_size: Option<u32>,
+    pub(crate) discard_oldest: Option<bool>,
+    pub(crate) _kind: PhantomData<K>,
 }
 
 impl MonitoredItemBuilder<DataChange<attributes::Value>> {
@@ -47,13 +48,15 @@ impl MonitoredItemBuilder<DataChange<attributes::Value>> {
 impl<K: MonitoredItemKind> MonitoredItemBuilder<K> {
     /// Sets attribute.
     ///
-    /// By default, monitored items emit [`DataValue`](crate::DataValue) of the appropriate subtype matching the given
-    /// attribute. If the attribute is set to [`ua::AttributeId::EVENTNOTIFIER_T`], they emit
-    /// `ua::Array<ua::Variant>` instead.
+    /// By default, monitored items emit [`DataValue`](crate::DataValue) of the appropriate subtype
+    /// matching the given attribute. If the attribute is set to [`ua::AttributeId::EVENTNOTIFIER_T`],
+    /// they emit `ua::Array<ua::Variant>` instead.
     ///
     /// Default value is [`ua::AttributeId::VALUE_T`].
     ///
     /// See [`Self::attribute_id()`] to set the attribute ID at runtime.
+    ///
+    /// See [`MonitoredItemCreateRequestBuilder::attribute()`].
     ///
     /// # Examples
     ///
@@ -113,7 +116,9 @@ impl<K: MonitoredItemKind> MonitoredItemBuilder<K> {
     ///
     /// Default value is [`ua::AttributeId::VALUE`].
     ///
-    /// See [`ua::MonitoredItemCreateRequest::with_attribute_id()`].
+    /// See:
+    ///   - [`MonitoredItemCreateRequestBuilder::attribute_id()`]
+    ///   - [`ua::MonitoredItemCreateRequest::with_attribute_id()`]
     ///
     /// # Examples
     ///
@@ -170,7 +175,9 @@ impl<K: MonitoredItemKind> MonitoredItemBuilder<K> {
     ///
     /// Default value is [`ua::MonitoringMode::REPORTING`].
     ///
-    /// See [`ua::MonitoredItemCreateRequest::with_monitoring_mode()`].
+    /// See:
+    ///   - [`MonitoredItemCreateRequestBuilder::monitoring_mode()`]
+    ///   - [`ua::MonitoredItemCreateRequest::with_monitoring_mode()`]
     #[must_use]
     pub fn monitoring_mode(mut self, monitoring_mode: ua::MonitoringMode) -> Self {
         self.monitoring_mode = Some(monitoring_mode);
@@ -181,7 +188,9 @@ impl<K: MonitoredItemKind> MonitoredItemBuilder<K> {
     ///
     /// Default value is 250.0 ms.
     ///
-    /// See [`ua::MonitoringParameters::with_sampling_interval()`].
+    /// See:
+    ///   - [`MonitoredItemCreateRequestBuilder::sampling_interval()`]
+    ///   - [`ua::MonitoredItemCreateRequest::with_sampling_interval()`]
     #[must_use]
     pub const fn sampling_interval(mut self, sampling_interval: Option<Duration>) -> Self {
         self.sampling_interval = Some(sampling_interval);
@@ -192,7 +201,9 @@ impl<K: MonitoredItemKind> MonitoredItemBuilder<K> {
     ///
     /// Default value is no filter.
     ///
-    /// See [`ua::MonitoringParameters::with_filter()`].
+    /// See:
+    ///   - [`MonitoredItemCreateRequestBuilder::filter()`]
+    ///   - [`ua::MonitoredItemCreateRequest::with_filter()`]
     #[must_use]
     pub fn filter(mut self, filter: impl MonitoringFilter) -> Self {
         self.filter = Some(Box::new(filter));
@@ -203,7 +214,9 @@ impl<K: MonitoredItemKind> MonitoredItemBuilder<K> {
     ///
     /// Default value is 1.
     ///
-    /// See [`ua::MonitoringParameters::with_queue_size()`].
+    /// See:
+    ///   - [`MonitoredItemCreateRequestBuilder::queue_size()`]
+    ///   - [`ua::MonitoredItemCreateRequest::with_queue_size()`]
     #[must_use]
     pub const fn queue_size(mut self, queue_size: u32) -> Self {
         self.queue_size = Some(queue_size);
@@ -214,7 +227,9 @@ impl<K: MonitoredItemKind> MonitoredItemBuilder<K> {
     ///
     /// Default value is `true`.
     ///
-    /// See [`ua::MonitoringParameters::with_discard_oldest()`].
+    /// See:
+    ///   - [`MonitoredItemCreateRequestBuilder::discard_oldest()`]
+    ///   - [`ua::MonitoredItemCreateRequest::with_discard_oldest()`]
     #[must_use]
     pub const fn discard_oldest(mut self, discard_oldest: bool) -> Self {
         self.discard_oldest = Some(discard_oldest);
@@ -233,12 +248,45 @@ impl<K: MonitoredItemKind> MonitoredItemBuilder<K> {
         self,
         subscription: &AsyncSubscription,
     ) -> Result<Vec<Result<(ua::MonitoredItemCreateResult, AsyncMonitoredItem<K>)>>> {
+        AsyncMonitoredItem::create(subscription, self.into()).await
+    }
+}
+
+/// Monitored item (with asynchronous API).
+#[derive(Debug)]
+pub struct AsyncMonitoredItem<K: MonitoredItemKind = DataChange<attributes::Value>> {
+    handle: MonitoredItemHandle,
+    rx: mpsc::Receiver<K::Value>,
+    _kind: PhantomData<K>,
+}
+
+impl<K: MonitoredItemKind> AsyncMonitoredItem<K> {
+    const fn new(handle: MonitoredItemHandle, rx: mpsc::Receiver<K::Value>) -> Self {
+        Self {
+            handle,
+            rx,
+            _kind: PhantomData,
+        }
+    }
+
+    /// Creates monitored items.
+    ///
+    /// This creates one or more new monitored items. Returns one result for each node ID.
+    ///
+    /// # Errors
+    ///
+    /// This fails when the entire request is not successful. Errors for individual node IDs are
+    /// returned as error elements inside the resulting list.
+    pub async fn create(
+        subscription: &AsyncSubscription,
+        request_builder: MonitoredItemCreateRequestBuilder<K>,
+    ) -> Result<Vec<Result<(ua::MonitoredItemCreateResult, AsyncMonitoredItem<K>)>>> {
         let Some(client) = &subscription.client().upgrade() else {
             return Err(Error::internal("client should not be dropped"));
         };
         let subscription_id = subscription.subscription_id();
 
-        let request = self.into_request(subscription_id);
+        let request = request_builder.build(subscription_id);
         let result_count = request.items_to_create().map_or(0, <[_]>::len);
         let mut rxs = Vec::with_capacity(result_count);
         let response = {
@@ -300,74 +348,11 @@ impl<K: MonitoredItemKind> MonitoredItemBuilder<K> {
                 let handle =
                     MonitoredItemHandle::new(client, subscription_id, result.monitored_item_id());
                 let monitored_item = AsyncMonitoredItem::new(handle, rx);
-
                 Ok((result, monitored_item))
             })
             .collect();
 
         Ok(results)
-    }
-
-    fn into_request(self, subscription_id: ua::SubscriptionId) -> ua::CreateMonitoredItemsRequest {
-        let Self {
-            node_ids,
-            attribute_id,
-            monitoring_mode,
-            sampling_interval,
-            filter,
-            queue_size,
-            discard_oldest,
-            _kind: _,
-        } = self;
-
-        let items_to_create = node_ids
-            .into_iter()
-            .map(|node_id| {
-                let mut request = ua::MonitoredItemCreateRequest::default()
-                    .with_node_id(&node_id)
-                    .with_attribute_id(&attribute_id);
-
-                if let Some(monitoring_mode) = monitoring_mode.as_ref() {
-                    request = request.with_monitoring_mode(monitoring_mode);
-                }
-                if let Some(&sampling_interval) = sampling_interval.as_ref() {
-                    request = request.with_sampling_interval(sampling_interval);
-                }
-                if let Some(filter) = filter.as_ref() {
-                    request = request.with_filter(filter);
-                }
-                if let Some(&queue_size) = queue_size.as_ref() {
-                    request = request.with_queue_size(queue_size);
-                }
-                if let Some(&discard_oldest) = discard_oldest.as_ref() {
-                    request = request.with_discard_oldest(discard_oldest);
-                }
-
-                request
-            })
-            .collect::<Vec<_>>();
-
-        ua::CreateMonitoredItemsRequest::init()
-            .with_subscription_id(subscription_id)
-            .with_items_to_create(&items_to_create)
-    }
-}
-
-/// Monitored item (with asynchronous API).
-#[derive(Debug)]
-pub struct AsyncMonitoredItem<K: MonitoredItemKind = DataChange<attributes::Value>> {
-    handle: MonitoredItemHandle,
-    rx: mpsc::Receiver<K::Value>,
-    _kind: PhantomData<K>,
-}
-
-impl<K: MonitoredItemKind> AsyncMonitoredItem<K> {
-    const fn new(handle: MonitoredItemHandle, rx: mpsc::Receiver<K::Value>) -> Self {
-        Self {
-            handle,
-            rx,
-            _kind: PhantomData,
-        }
     }
 
     /// Waits for next value from server.