+++ src/monitored_item.rs
@@ -5,13 +5,8 @@
         reason = "Some methods are only used when this feature is enabled."
     )
 )]
-#![cfg_attr(
-    not(feature = "tokio"),
-    expect(
-        unreachable_pub,
-        reason = "Some types/methods only need to be public when this features is enabled."
-    )
-)]
+
+mod create_request_builder;
 
 // TODO: Remove pub(crate).
 pub(crate) mod create_monitored_items;
@@ -24,6 +19,8 @@ use std::{
 };
 
 use crate::{attributes, ua, Attribute, DataType as _, DataValue, Error, Result};
+
+pub use self::create_request_builder::MonitoredItemCreateRequestBuilder;
 
 /// Handle for a single monitored item.
 ///
@@ -121,7 +118,8 @@ impl Drop for MonitoredItemHandle {
 
 /// Value emitted from monitored item notification.
 ///
-/// This depends on the attribute ID passed to [`MonitoredItemBuilder::attribute_id()`](crate::MonitoredItemBuilder::attribute_id).
+/// This depends on the attribute ID passed to
+/// [`MonitoredItemCreateRequestBuilder::attribute_id()`](crate::MonitoredItemCreateRequestBuilder::attribute_id).
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct MonitoredItemValue(MonitoredItemValueInner);
 
@@ -242,7 +240,7 @@ impl MonitoredItemKind for Unknown {
 /// This is implemented for all attributes except [`ua::AttributeId::EVENTNOTIFIER_T`].
 trait DataChangeAttribute: Attribute {}
 
-/// Attribute for [`MonitoredItemBuilder::attribute()`](crate::MonitoredItemBuilder::attribute).
+/// Attribute for [`MonitoredItemCreateRequestBuilder::attribute()`](crate::MonitoredItemCreateRequestBuilder::attribute).
 pub trait MonitoredItemAttribute: Attribute {
     /// Matching [`MonitoredItemKind`] implementation for attribute.
     type Kind: MonitoredItemKind;